Refactor Drawer
This PR tries to simplify the `Drawer` definitions by consolidating as
much code as possible into a backend-agnostic file. One major change is
to stop `clear` writing directly to the underlying surface. Instead,
everything is written to the RecordingSurface.

This has a number of knock-on impacts:
- Bar borders need to redrawn each time the bar is drawn
- Widget screenshots need to access the recording surface
