+++ libqtile/backend/wayland/drawer.py
@@ -1,3 +1,23 @@
+# Copyright (c) 2022 m-col
+# Copyright (c) 2023 elParaguayo
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
 from __future__ import annotations
 
 from typing import TYPE_CHECKING
@@ -5,15 +25,15 @@ from typing import TYPE_CHECKING
 import cairocffi
 from wlroots.util.region import PixmanRegion32
 
-from libqtile.backend import base
+from libqtile import utils
+from libqtile.backend import drawer
 
 if TYPE_CHECKING:
     from libqtile.backend.wayland.window import Internal
     from libqtile.core.manager import Qtile
-    from libqtile.utils import ColorsType
 
 
-class Drawer(base.Drawer):
+class Drawer(drawer.Drawer):
     """
     A helper class for drawing and text layout.
 
@@ -22,7 +42,7 @@ class Drawer(base.Drawer):
     """
 
     def __init__(self, qtile: Qtile, win: Internal, width: int, height: int):
-        base.Drawer.__init__(self, qtile, win, width, height)
+        drawer.Drawer.__init__(self, qtile, win, width, height)
 
     def _draw(
         self,
@@ -66,12 +86,3 @@ class Drawer(base.Drawer):
         # TODO: do we really need to `set_buffer` here? would be good to just set damage
         self._win._scene_buffer.set_buffer_with_damage(self._win.wlr_buffer, damage)
         damage.fini()
-
-    def clear(self, colour: ColorsType) -> None:
-        # Draw background straight to ImageSurface
-        ctx = cairocffi.Context(self.surface)
-        ctx.save()
-        ctx.set_operator(cairocffi.OPERATOR_SOURCE)
-        self.set_source_rgb(colour, ctx=ctx)
-        ctx.paint()
-        ctx.restore()
